+++ src/components/inputs/DatePicker/tests/DatePicker.test.tsx
@@ -1,0 +1,16 @@
+import React from "react";
+import { render } from "@testing-library/react";
+
+import DatePicker from "../DatePicker";
+
+describe("DatePicker", () => {
+  test("Renders the DatePicker component", () => {
+    render(
+      <DatePicker
+        onChange={(date: Date): void => {
+          console.log(Date);
+        }}
+      />
+    );
+  });
+});